put frollmedian in DT structure
