+++ NAMESPACE
@@ -54,6 +54,7 @@ export(frollsum)
 export(frollmax)
 export(frollmin)
 export(frollprod)
+export(frollmedian)
 export(frollapply)
 export(frolladapt)
 export(nafill)
