+++ man/froll.Rd
@@ -9,11 +9,13 @@
 \alias{frollmax}
 \alias{frollmin}
 \alias{frollprod}
+\alias{frollmedian}
 \alias{rollmean}
 \alias{rollsum}
 \alias{rollmax}
 \alias{rollmin}
 \alias{rollprod}
+\alias{rollmedian}
 \title{Rolling functions}
 \description{
   Fast rolling functions to calculate aggregates on sliding windows. For user-defined rolling function see \code{\link{frollapply}}.
@@ -28,6 +30,8 @@
   frollmin(x, n, fill=NA, algo=c("fast","exact"), align=c("right","left","center"),
     na.rm=FALSE, has.nf=NA, adaptive=FALSE, partial=FALSE, give.names=FALSE, hasNA)
   frollprod(x, n, fill=NA, algo=c("fast","exact"), align=c("right","left","center"),
+    na.rm=FALSE, has.nf=NA, adaptive=FALSE, partial=FALSE, give.names=FALSE, hasNA)
+  frollmedian(x, n, fill=NA, algo=c("fast","exact"), align=c("right","left","center"),
     na.rm=FALSE, has.nf=NA, adaptive=FALSE, partial=FALSE, give.names=FALSE, hasNA)
 }
 \arguments{
@@ -72,11 +76,11 @@
   }
 }
 \section{Implementation}{
-  Each rolling function has 4 different implementations. First factor that decides which implementation is being used is \code{adaptive} argument, see setion below for details. Then for each of those two algorithms (adaptive \code{TRUE/FALSE}) there are two \code{algo} argument values.
+  Most of the rolling functions have 4 different implementations. First factor that decides which implementation is being used is \code{adaptive} argument, see setion below for details. Then for each of those two algorithms (adaptive \code{TRUE/FALSE}) there are usually two \code{algo} argument values.
   \itemize{
-    \item{ \code{algo="fast"} uses \emph{"on-line"}, single pass, algorithm.
+    \item{ \code{algo="fast"} uses \emph{"online"}, single pass, algorithm.
     \itemize{
-      \item{ \emph{max} and \emph{min} rolling function will not do only a single pass but, on average \code{length(x)/n}, nested loops will be computed. The bigger the window the bigger advantage over algo \emph{exact} which computes \code{length(x)} nested loops. Note that \emph{exact} uses multiple CPUs so for a small window size and many CPUs it is possible it will be actually faster than \emph{fast} but in those cases elapsed timings will likely be far below a single second. }
+      \item{ \emph{max} and \emph{min} rolling function will not do only a single pass but, on average \code{length(x)/n}, nested loops will be computed. The bigger the window the bigger advantage over algo \emph{exact} which computes \code{length(x)} nested loops. Note that \emph{exact} uses multiple CPUs so for a small window size and many CPUs it is possible it will be actually faster than \emph{fast} but in those cases elapsed timings will likely be close to zero. }
       \item{ Not all functions have \emph{fast} implementation available. As of now \emph{max} and \emph{min} in case of \code{adaptive=TRUE} do not not have \emph{fast} implementation, therefore it will automatically fall back to \emph{exact} implementation. \code{datatable.verbose} option can be used to check that. }
     }}
     \item{ \code{algo="exact"} will make rolling functions to use a more computationally-intensive algorithm. For each observation from input vector it will compute a function on a window from scratch (complexity \eqn{O(n^2)}).
@@ -85,6 +89,7 @@
       \item{ In case of \emph{mean} (and possibly other functions in future), it will additionally make extra pass to perform floating point error correction. Error corrections might not be truly exact on some platforms (like Windows) when using multiple threads. }
     }}
   }
+  Rolling median is implemented according to a novel algorithm described by \emph{Jukka Suomela} in his paper \emph{Median Filtering is Equivalent to Sorting (2014)}. See references section for the link. Implementation here is extended to support arbitrary length of input and as well an even window size. Despite extensive validation of results this function should be considered experimental.
 }
 \section{Adaptive rolling functions}{
   Adaptive rolling functions are a special case where each observation has its own corresponding rolling window width. Therefore values passed to \code{n} argument must be series corresponding to observations in \code{x}. If multiple windows is meant to be computed then a list of integer vectors is expected; each list element must be an integer vector of window size corresponding to observations in \code{x}; see Examples. Due to the logic or implementation of adaptive rolling functions, the following restrictions apply
@@ -104,7 +109,7 @@
     \item{ \code{fill} accepts only constant values. No support for \emph{na.locf} or other functions. }
     \item{ \code{align} defaults to \code{"right"}. }
     \item{ \code{na.rm} is respected, and other functions are not needed when input contains \code{NA}. }
-    \item{ integers and logical are always coerced to double. }
+    \item{ integers and logical are always coerced to double before computation. }
     \item{ when \code{adaptive=FALSE} (default), then \code{n} must be a numeric vector. List is not accepted. }
     \item{ when \code{adaptive=TRUE}, then \code{n} must be vector of length equal to \code{nrow(x)}, or list of such vectors. }
   }
@@ -129,6 +134,7 @@ frollsum(d, 3:4)
 frollmax(d, 3:4)
 frollmin(d, 3:4)
 frollprod(d, 3:4)
+frollmedian(d, 3:4)
 
 # partial=TRUE
 x = 1:6/2
@@ -184,6 +190,6 @@ sapply(errs, format, scientific=FALSE) # roundoff
   \code{\link{frollapply}}, \code{\link{frolladapt}}, \code{\link{shift}}, \code{\link{data.table}}, \code{\link{setDTthreads}}
 }
 \references{
-  \href{https://en.wikipedia.org/wiki/Round-off_error}{Round-off error}
+  \href{https://en.wikipedia.org/wiki/Round-off_error}{Round-off error}, \href{https://arxiv.org/abs/1406.1717}{"Median Filtering is Equivalent to Sorting" by Jukka Suomela}
 }
 \keyword{ data }
